size option in product creation

+++ api_clothe-store/src/app/services/product.service.js
@@ -1,5 +1,5 @@
 import { consulta } from "../database/connection.js"
-
+import connection from "../database/connection.js"
 
 class productService{
 
@@ -7,6 +7,7 @@ class productService{
         const sql = "SELECT * FROM clothes"
         return consulta(sql, '' ,'Not Found!')
     }
+    
 
     findById(id){
         const sql = "SELECT * FROM clothes WHERE id=?"
@@ -32,9 +33,48 @@ class productService{
 
 
     posting(dados){
-        let sql = "INSERT INTO clothes SET ?"
-        return consulta(sql, dados, 'Not Created!')
+
+        return new Promise(async (resolve, reject) =>{
+
+            if(!dados){ 
+                return reject({message: 'Data missing!'})
+            }
+            
+            const newSizes = dados.sizes.split(',').map(Number)
+            const {sizes, ...transformedDados} = dados
+
+
+            let sql = "INSERT INTO clothes (name, image, info, isFavorite, sexo, section, price, isBestseller, quantity) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)"
+            
+
+            const [res_productCreated] = await connection.promise().execute(sql, [
+                transformedDados.name,
+                transformedDados.image,
+                transformedDados.info,
+                transformedDados.isFavorite,
+                transformedDados.sexo,
+                transformedDados.section,
+                transformedDados.price,
+                transformedDados.isBestseller,
+                transformedDados.quantity,
+            ])
+
+            const insertId = res_productCreated.insertId
+            
+            const values = newSizes.map(() => '(?, ?)').join(', ')
+            const params = newSizes.flatMap(size => [insertId, size])
+
+            const sql2 = `INSERT INTO products_size (product_id, size_id) VALUES ${values}`
+
+            await connection.promise().execute(sql2, params)
+            
+            resolve({message: 'Product Created!'})
+
+        })
+        
+
     }
+
 
     editing(id, dados){
         let sql = "UPDATE clothes SET ? WHERE id=?"
@@ -49,9 +89,25 @@ class productService{
         return consulta(sql, [valor, id], 'Erro in Edit Product!')
     }
 
+
     delete(id){
-        const sql = 'DELETE FROM clothes WHERE id=?'
-        return consulta(sql, id, 'Erro in Delete Product!')
+
+        return new Promise(async (resolve, reject) =>{
+
+            if(!id) return reject({message: 'Id missing in delete product service!'})
+
+            const sql = 'DELETE FROM clothes WHERE id = ?'
+            await connection.promise().execute(sql, [id])
+            console.log("✅ Product deleted")
+
+            const sql2 = 'DELETE FROM products_size WHERE product_id = ?'
+            await connection.promise().execute(sql2, [id])
+            console.log("✅ product size deleted") 
+
+            return resolve({message: 'Product Deleted!'})
+
+        })
+        
     }
 
 }
